Beetlebot/requirements.txt
discord.py==2.2.2
openai==0.27.5